requirements.txt
sqlalchemy==1.4.46
pandas==1.5.3
python-dotenv==1.0.0
